docs/하은,윤민/requirements_eval_pipeline.txt
openai>=1.40.0
python-dotenv>=1.0.1
chromadb==0.5.23
numpy>=1.26.0
rank-bm25>=0.2.2
langchain-core>=0.3.0
langchain-openai>=0.2.0
streamlit>=1.36.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
olefile>=0.47
python-docx>=1.1.2
